+++ spack-scripting/scripting/cmd/manager_cmds/external.py
@@ -9,7 +9,7 @@ import os
 import re
 from datetime import datetime
 
-from manager_utils import base_extension
+from manager_utils import base_extension, pruned_spec_string
 
 import llnl.util.tty as tty
 
@@ -98,7 +98,7 @@ def add_include_entry(env, inc, prepend=True):
 
 def create_external_detected_spec(env, spec):
     view = _get_first_view_containing_spec(env, spec)
-    pruned_spec = _well_posed_spec_string_minus_dev_path(spec)
+    pruned_spec = pruned_spec_string(spec)
     prefix = view.get_projection_for_spec(spec)
     return spack.detection.DetectedPackage(Spec.from_detection(pruned_spec), prefix)
 
@@ -136,23 +136,6 @@ def create_yaml_from_detected_externals(ext_dict):
         formatted_dict[name] = config
 
     return syaml.syaml_dict({"packages": formatted_dict})
-
-
-def _well_posed_spec_string_minus_dev_path(spec):
-    full_spec = spec.format("{name}{@version}{%compiler}{variants}{arch=architecture}")
-    spec_components = full_spec.split(" ")
-    variants_to_omit = ("dev_path=", "patches=")
-
-    def filter_func(entry):
-        for v in variants_to_omit:
-            if v in entry:
-                return False
-        return True
-
-    pruned_components = list(filter(filter_func, spec_components))
-
-    pruned_spec = " ".join(pruned_components)
-    return pruned_spec
 
 
 def _get_first_view_containing_spec(env, spec):
